backend/requirements.txt
django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
django_filter==23.1